feat(logserver-spark): Разделение по чистой архитектуре
Ref: https://www.youtube.com/watch?v=vXVoGkSmWjo
Ref: https://github.com/tupol/spark-utils
